[Add]: 재생컴포넌트 분리
